Change token_id type from bigint to smallint in god_metadatas table

+++ database/create-table/god_metadatas.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS "public"."god_metadatas" (
-  "token_id" bigint NOT NULL,
+  "token_id" smallint NOT NULL,
   "type" "text" NOT NULL,
   "gender" "text" NOT NULL,
   "parts" "jsonb" NOT NULL,
